Agent/Step 5/requirements.txt
python-dotenv
requests
livekit-agents[openai,silero,turn-detector]
livekit-plugins-openai
livekit-plugins-silero
livekit-api